super_resolution/pyproject.toml
[project]
name = "super-resolution"
version = "0.1.0"
description = "Super Resolution using EDSR model from Hugging Face"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "super-image>=0.1.0",
    "huggingface-hub<0.14",
    "pillow>=11.2.1",
    "typer>=0.15.3",
]

[project.scripts]
super-resolve = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."] 